[snomed] merge refset.core into snomed.datastore
