+++ snomed/com.b2international.snowowl.snomed.core.feature/feature.xml
@@ -73,13 +73,6 @@ Visit us at http://b2i.sg
          unpack="false"/>
 
    <plugin
-         id="com.b2international.snowowl.snomed.refset.core"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
-   <plugin
          id="com.b2international.snowowl.snomed.refset.model"
          download-size="0"
          install-size="0"
