+++ snomed/com.b2international.snowowl.snomed.datastore.server/META-INF/MANIFEST.MF
@@ -11,7 +11,6 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.9.0",
  com.b2international.snowowl.snomed.datastore;visibility:=reexport,
  com.b2international.snowowl.rpc,
  com.b2international.snowowl.snomed.scg,
- com.b2international.snowowl.snomed.refset.core,
  com.b2international.snowowl.server.console
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
