+++ snomed/com.b2international.snowowl.snomed.datastore/META-INF/MANIFEST.MF
@@ -21,11 +21,13 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.9.0",
  com.b2international.snowowl.snomed.mrcm.model;visibility:=reexport,
  com.b2international.snowowl.eventbus,
  com.b2international.snowowl.snomed.ecl,
+ com.b2international.snowowl.dependencies,
  org.apache.httpcomponents.httpclient;bundle-version="4.2.6",
  org.apache.httpcomponents.httpcore;bundle-version="4.2.5",
  com.b2international.mapdb;bundle-version="[3.0.5,3.0.6)",
  com.fasterxml.jackson.dataformat.jackson-dataformat-csv;bundle-version="2.8.6",
- com.b2international.snowowl.core
+ com.b2international.snowowl.core,
+ org.eclipse.core.resources
 Export-Package: com.b2international.snowowl.snomed.core.domain,
  com.b2international.snowowl.snomed.core.domain.constraint,
  com.b2international.snowowl.snomed.core.domain.refset,
@@ -35,6 +37,7 @@ Export-Package: com.b2international.snowowl.snomed.core.domain,
  com.b2international.snowowl.snomed.core.mrcm,
  com.b2international.snowowl.snomed.core.mrcm.io,
  com.b2international.snowowl.snomed.core.preference,
+ com.b2international.snowowl.snomed.core.refset.automap,
  com.b2international.snowowl.snomed.core.store,
  com.b2international.snowowl.snomed.core.tree,
  com.b2international.snowowl.snomed.datastore,
