+++ snomed/com.b2international.snowowl.snomed.importer.rf2/META-INF/MANIFEST.MF
@@ -28,5 +28,4 @@ Export-Package: com.b2international.snowowl.snomed.importer.rf2,com.b2
  f2.validation
 Service-Component: OSGI-INF/commandprovider.xml
 Import-Package: com.b2international.snowowl.api.impl.codesystem.domain,
- com.b2international.snowowl.snomed.refset.core.automap,
  org.slf4j;version="1.7.2"
