--- snomed/com.b2international.snowowl.snomed.refset.core/META-INF/MANIFEST.MF
@@ -1,21 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Snow Owl SNOMED CT Reference Set Core
-Bundle-SymbolicName: com.b2international.snowowl.snomed.refset.core;singleton:=true
-Bundle-Version: 6.4.0.qualifier
-Bundle-Activator: com.b2international.snowowl.snomed.refset.core.SnomedRefsetCoreActivator
-Bundle-Vendor: B2i Healthcare
-Require-Bundle: org.eclipse.core.runtime;bundle-version="3.9.0",
- org.eclipse.core.resources;bundle-version="3.8.0",
- xstream;bundle-version="1.4.7",
- com.b2international.snowowl.snomed.datastore,
- com.b2international.snowowl.snomed.semanticengine.simpleast,
- com.b2international.snowowl.dependencies
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-ActivationPolicy: lazy
-Export-Package: com.b2international.snowowl.snomed.refset.clone,
- com.b2international.snowowl.snomed.refset.core,
- com.b2international.snowowl.snomed.refset.core.automap,
- com.b2international.snowowl.snomed.refset.core.compare,
- com.b2international.snowowl.snomed.refset.derivation
-Import-Package: org.slf4j;version="1.7.2"
