--- snomed/com.b2international.snowowl.snomed.refset.core/plugin.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension-point id="automapService" name="automapService" schema="schema/automapService.exsd"/>
-
-</plugin>
