--- snomed/com.b2international.snowowl.snomed.refset.core/pom.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>com.b2international.snowowl</groupId>
-    <artifactId>snomed-parent</artifactId>
-    <version>6.4.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>com.b2international.snowowl.snomed.refset.core</artifactId>
-  <packaging>eclipse-plugin</packaging>
-</project>
